fix(terminal): restore state-machine serial ordering for OSC/CSI responses
Fixes gh/termenv leaking OSC color responses into shell prompt.

Root cause: upstream Java termux-app calls mSession.write() synchronously
inside doOsc(), ensuring immediate ordered responses. The Rust rewrite
buffered terminal responses into pending_events and only drained them at
the end of process_bytes(), causing CSI responses to appear before OSC.

Changes:
- Add drain_pending_events() in PerformHandler after each csi_dispatch
  and osc_dispatch to preserve input-processing order.
- Fix VTE parser ESC_OSC_ESC state: ST-terminated queries reply with ST,
  BEL-terminated with BEL (commit 7f46a5e2 follow-up).
- Fix u16 multiply overflow in generate_color_report() using u32 intermediate.
- Optimize pending_events locking: replace std::sync::Mutex with
  parking_lot::Mutex + AtomicBool fast-path to recover ANSI throughput
  from ~41 MB/s back to ~46.5 MB/s (-10% vs pre-fix buggy baseline).
- Add regression tests for OSC/CSI response ordering (BEL, ST, bulk).
- Add validate_termux_prefix() firewall with 4 unit tests to prevent
  dynamic PREFIX regressions.

Refs: commit 5121e9e0 (PREFIX), commit 7f46a5e2 (OSC terminator)
Performance: ANSI escape throughput 46.5 MB/s (was 51.9 MB/s pre-fix)

+++ terminal-emulator/src/main/rust/src/terminal/colors.rs
@@ -249,9 +249,10 @@ impl TerminalColors {
             return String::new();
         }
         let color = self.current_colors[index];
-        let r = (((color >> 16) & 0xff) as u16 * 65535) / 255;
-        let g = (((color >> 8) & 0xff) as u16 * 65535) / 255;
-        let b = ((color & 0xff) as u16 * 65535) / 255;
+        // 使用 u32 中间计算避免 u16 溢出：255 * 65535 = 16_711_425 > u16::MAX
+        let r = ((((color >> 16) & 0xff) as u32 * 65535) / 255) as u16;
+        let g = ((((color >> 8) & 0xff) as u32 * 65535) / 255) as u16;
+        let b = (((color & 0xff) as u32 * 65535) / 255) as u16;
         format!("rgb:{:04x}/{:04x}/{:04x}", r, g, b)
     }
 }
